+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * `add_weekly_avg` computes the weekly average of a given variable
+* added `summarise_covid_doubling`, which computes observed time to 
+  double covid19 cases